+++ src/main/java/com/esmods/keepersofthestonestwo/item/FilthArmorItem.java
@@ -58,7 +58,7 @@ public abstract class FilthArmorItem extends ArmorItem {
 		public void inventoryTick(ItemStack itemstack, Level world, Entity entity, int slot, boolean selected) {
 			super.inventoryTick(itemstack, world, entity, slot, selected);
 			if (entity instanceof Player player && Iterables.contains(player.getArmorSlots(), itemstack)) {
-				RemoveForbiddenItemProcedure.execute(entity, itemstack);
+				RemoveForbiddenItemProcedure.execute(world, entity, itemstack);
 			}
 		}
 	}
@@ -72,7 +72,7 @@ public abstract class FilthArmorItem extends ArmorItem {
 		public void inventoryTick(ItemStack itemstack, Level world, Entity entity, int slot, boolean selected) {
 			super.inventoryTick(itemstack, world, entity, slot, selected);
 			if (entity instanceof Player player && Iterables.contains(player.getArmorSlots(), itemstack)) {
-				RemoveForbiddenItemProcedure.execute(entity, itemstack);
+				RemoveForbiddenItemProcedure.execute(world, entity, itemstack);
 			}
 		}
 	}
@@ -86,7 +86,7 @@ public abstract class FilthArmorItem extends ArmorItem {
 		public void inventoryTick(ItemStack itemstack, Level world, Entity entity, int slot, boolean selected) {
 			super.inventoryTick(itemstack, world, entity, slot, selected);
 			if (entity instanceof Player player && Iterables.contains(player.getArmorSlots(), itemstack)) {
-				RemoveForbiddenItemProcedure.execute(entity, itemstack);
+				RemoveForbiddenItemProcedure.execute(world, entity, itemstack);
 			}
 		}
 	}
@@ -100,7 +100,7 @@ public abstract class FilthArmorItem extends ArmorItem {
 		public void inventoryTick(ItemStack itemstack, Level world, Entity entity, int slot, boolean selected) {
 			super.inventoryTick(itemstack, world, entity, slot, selected);
 			if (entity instanceof Player player && Iterables.contains(player.getArmorSlots(), itemstack)) {
-				RemoveForbiddenItemProcedure.execute(entity, itemstack);
+				RemoveForbiddenItemProcedure.execute(world, entity, itemstack);
 			}
 		}
 	}