+++ src/main/java/com/esmods/keepersofthestonestwo/item/HeatKnifeItem.java
@@ -65,6 +65,6 @@ public class HeatKnifeItem extends SwordItem {
 	@Override
 	public void inventoryTick(ItemStack itemstack, Level world, Entity entity, int slot, boolean selected) {
 		super.inventoryTick(itemstack, world, entity, slot, selected);
-		RemoveForbiddenItemProcedure.execute(entity, itemstack);
+		RemoveForbiddenItemProcedure.execute(world, entity, itemstack);
 	}
 }