+++ src/main/java/com/esmods/keepersofthestonestwo/item/SandArmorItem.java
@@ -59,7 +59,7 @@ public abstract class SandArmorItem extends ArmorItem {
 		public void inventoryTick(ItemStack itemstack, Level world, Entity entity, int slot, boolean selected) {
 			super.inventoryTick(itemstack, world, entity, slot, selected);
 			if (entity instanceof Player player && Iterables.contains(player.getArmorSlots(), itemstack)) {
-				RemoveForbiddenItemProcedure.execute(entity, itemstack);
+				RemoveForbiddenItemProcedure.execute(world, entity, itemstack);
 			}
 		}
 	}
@@ -87,7 +87,7 @@ public abstract class SandArmorItem extends ArmorItem {
 		public void inventoryTick(ItemStack itemstack, Level world, Entity entity, int slot, boolean selected) {
 			super.inventoryTick(itemstack, world, entity, slot, selected);
 			if (entity instanceof Player player && Iterables.contains(player.getArmorSlots(), itemstack)) {
-				RemoveForbiddenItemProcedure.execute(entity, itemstack);
+				RemoveForbiddenItemProcedure.execute(world, entity, itemstack);
 			}
 		}
 	}
@@ -101,7 +101,7 @@ public abstract class SandArmorItem extends ArmorItem {
 		public void inventoryTick(ItemStack itemstack, Level world, Entity entity, int slot, boolean selected) {
 			super.inventoryTick(itemstack, world, entity, slot, selected);
 			if (entity instanceof Player player && Iterables.contains(player.getArmorSlots(), itemstack)) {
-				RemoveForbiddenItemProcedure.execute(entity, itemstack);
+				RemoveForbiddenItemProcedure.execute(world, entity, itemstack);
 			}
 		}
 	}