+++ src/main/java/com/esmods/keepersofthestonestwo/item/ShockwaveGunItem.java
@@ -45,7 +45,7 @@ public class ShockwaveGunItem extends Item {
 	@Override
 	public void inventoryTick(ItemStack itemstack, Level world, Entity entity, int slot, boolean selected) {
 		super.inventoryTick(itemstack, world, entity, slot, selected);
-		RemoveForbiddenItemProcedure.execute(entity, itemstack);
+		RemoveForbiddenItemProcedure.execute(world, entity, itemstack);
 	}
 
 	@Override