+++ src/main/java/com/esmods/keepersofthestonestwo/procedures/CursedKeeperAirProcedure.java
@@ -13,19 +13,23 @@ public class CursedKeeperAirProcedure {
 			return;
 		if (entity instanceof LivingEntity _entity && !_entity.level().isClientSide())
 			_entity.addEffect(new MobEffectInstance(MobEffects.MOVEMENT_SLOWDOWN, 3, 5, false, false));
-		entity.getPersistentData().putDouble("IA", (entity.getPersistentData().getDouble("IA") + 1));
-		if (entity.getPersistentData().getDouble("IA") == 15) {
+		if (entity instanceof CursedKeeperEntity _datEntSetI)
+			_datEntSetI.getEntityData().set(CursedKeeperEntity.DATA_IA, (int) ((entity instanceof CursedKeeperEntity _datEntI ? _datEntI.getEntityData().get(CursedKeeperEntity.DATA_IA) : 0) + 1));
+		if ((entity instanceof CursedKeeperEntity _datEntI ? _datEntI.getEntityData().get(CursedKeeperEntity.DATA_IA) : 0) == 15) {
 			if (entity instanceof CursedKeeperEntity _datEntSetL)
 				_datEntSetL.getEntityData().set(CursedKeeperEntity.DATA_stage_two_anim_sync, false);
 			if (entity instanceof CursedKeeperEntity _datEntSetI)
 				_datEntSetI.getEntityData().set(CursedKeeperEntity.DATA_attack_anim_sync, 3);
 		}
-		if (entity.getPersistentData().getDouble("IA") == 44) {
-			entity.getPersistentData().putDouble("windShield", 300);
+		if ((entity instanceof CursedKeeperEntity _datEntI ? _datEntI.getEntityData().get(CursedKeeperEntity.DATA_IA) : 0) == 44) {
+			if (entity instanceof CursedKeeperEntity _datEntSetI)
+				_datEntSetI.getEntityData().set(CursedKeeperEntity.DATA_windShield, 300);
 		}
-		if (entity.getPersistentData().getDouble("IA") == 344) {
-			entity.getPersistentData().putDouble("IA", 0);
-			entity.getPersistentData().putString("State", "Idle");
+		if ((entity instanceof CursedKeeperEntity _datEntI ? _datEntI.getEntityData().get(CursedKeeperEntity.DATA_IA) : 0) == 344) {
+			if (entity instanceof CursedKeeperEntity _datEntSetI)
+				_datEntSetI.getEntityData().set(CursedKeeperEntity.DATA_IA, 0);
+			if (entity instanceof CursedKeeperEntity _datEntSetS)
+				_datEntSetS.getEntityData().set(CursedKeeperEntity.DATA_State, "Idle");
 			if (entity instanceof CursedKeeperEntity _datEntSetI)
 				_datEntSetI.getEntityData().set(CursedKeeperEntity.DATA_attack_anim_sync, 0);
 			if (entity instanceof CursedKeeperEntity _datEntSetL)