+++ CHANGELOG.md
@@ -6,12 +6,13 @@
 - The stun effect will no longer constantly emit sound
 - Stun effect now lowers the player's stats
 - Admins now see the results of the pw command
+- Improved AI of mobs
 
 **API changes**
 - *N/A*
 
 **Bugfixes**
-- [Bugfix] 
+- [Bugfix] Fixed a bug that could cause the cursed keeper to misbehave in battle
 
 **Known issues**
 - *N/A*